requirements.txt
requests>=2.31.0
reportlab>=3.6.12
PyPDF2>=3.0.1
pdfminer.six>=20221105
pdf2image>=1.17.0
pillow>=10.0.0
pytesseract>=0.3.10

